+++ src/main/java/com/marcoswolf/crm/reparos/infrastructure/repositories/ReparoRepository.java
@@ -5,13 +5,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ReparoRepository extends JpaRepository<Reparo, Integer> {
+public interface ReparoRepository extends JpaRepository<Reparo, Long> {
     // Reparo
     List<Reparo> findByStatus_NomeContainingIgnoreCase(String nomeStatus);
     // Cliente
-    List<Reparo> findByEquipamento_Cliente_Id(Integer id);
+    List<Reparo> findByEquipamento_Cliente_Id(Long id);
     // Equipamento
-    List<Reparo> findByEquipamento_Id(Integer id);
+    List<Reparo> findByEquipamento_Id(Long id);
     // StatusReparo
-    List<Reparo> findByStatus_Id(Integer id);
+    List<Reparo> findByStatus_Id(Long id);
 }
