+++ src/main/java/com/marcoswolf/crm/reparos/infrastructure/repositories/TipoEquipamentoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TipoEquipamentoRepository extends JpaRepository<TipoEquipamento, Integer> {
+public interface TipoEquipamentoRepository extends JpaRepository<TipoEquipamento, Long> {
     // TipoEquipamento
     List<TipoEquipamento> findByNomeContainingIgnoreCase(String nome);
 }
